Move convex_hull from /util to /systems/robotInterfaces
It's only used from within controllers, and shouldn't be used elsewhere.
We mark it private, since it's only used from .cc files, also.
