+++ drake/systems/robotInterfaces/BUILD.bazel
@@ -1,6 +1,10 @@
 # -*- python -*-
 
-load("//tools:drake.bzl", "drake_cc_library")
+load(
+    "//tools/skylark:drake_cc.bzl",
+    "drake_cc_googletest",
+    "drake_cc_library",
+)
 load("//tools/lint:lint.bzl", "add_lint_tests")
 
 package(default_visibility = ["//visibility:public"])
@@ -20,6 +24,7 @@ drake_cc_library(
     hdrs = ["QPLocomotionPlan.h"],
     deps = [
         ":body_motion_data",
+        ":convex_hull",
         ":side",
         "//drake/common:essential",
         "//drake/common:unused",
@@ -34,7 +39,6 @@ drake_cc_library(
         "//drake/multibody:rigid_body_tree",
         "//drake/systems/controllers:zmp_util",
         "//drake/util",
-        "//drake/util:convex_hull",
         "//drake/util:lcm_util",
     ],
 )
@@ -44,6 +48,25 @@ drake_cc_library(
     srcs = ["Side.cpp"],
     hdrs = ["Side.h"],
     deps = [],
+)
+
+drake_cc_library(
+    name = "convex_hull",
+    srcs = ["convex_hull.cc"],
+    hdrs = ["convex_hull.h"],
+    visibility = [],
+    deps = [
+        "//drake/common:essential",
+    ],
+)
+
+# === test/ ===
+
+drake_cc_googletest(
+    name = "convex_hull_test",
+    deps = [
+        ":convex_hull",
+    ],
 )
 
 add_lint_tests()
