+++ drake/util/BUILD.bazel
@@ -22,15 +22,6 @@ drake_cc_library(
 )
 
 drake_cc_library(
-    name = "convex_hull",
-    srcs = ["convexHull.cpp"],
-    hdrs = ["convexHull.h"],
-    deps = [
-        "//drake/common:essential",
-    ],
-)
-
-drake_cc_library(
     name = "lcm_util",
     srcs = ["lcmUtil.cpp"],
     hdrs = ["lcmUtil.h"],
@@ -48,14 +39,6 @@ drake_cc_library(
 )
 
 # === test/ ===
-
-drake_cc_googletest(
-    name = "testConvexHull",
-    srcs = ["test/testConvexHull.cpp"],
-    deps = [
-        ":convex_hull",
-    ],
-)
 
 drake_cc_googletest(
     name = "testDrakeGeometryUtil",
