+++ tools/install/libdrake/build_components.bzl
@@ -325,7 +325,6 @@ LIBDRAKE_COMPONENTS = [
     "//drake/systems/trajectory_optimization:direct_collocation",
     "//drake/systems/trajectory_optimization:direct_transcription",
     "//drake/systems/trajectory_optimization:multiple_shooting",
-    "//drake/util:convex_hull",
     "//drake/util:lcm_util",
     "//drake/util:util",
 ]
